Implement swap table

+++ src/vm/swap.h
@@ -2,6 +2,6 @@
 #include <block.h>
 
 void swap_init ();
-uint32_t find_empty_swap_slot (); 
+size_t find_empty_swap_slot (); 
 bool swap_out (struct frame_table_entry *);
 bool swap_in (struct sPage_table_entry *);
